+++ WWPTools.WpfUI/obj/Release/net8.0-windows/WWPTools.WpfUI.AssemblyInfo.cs
@@ -11,7 +11,7 @@ using System;
 using System.Reflection;
 
 [assembly: System.Reflection.AssemblyCompanyAttribute("WWPTools.WpfUI")]
-[assembly: System.Reflection.AssemblyConfigurationAttribute("Debug")]
+[assembly: System.Reflection.AssemblyConfigurationAttribute("Release")]
 [assembly: System.Reflection.AssemblyFileVersionAttribute("1.0.0.0")]
 [assembly: System.Reflection.AssemblyInformationalVersionAttribute("1.0.0+1e8e27f8ec9a4e5297b5bcb59827e0b43de8f629")]
 [assembly: System.Reflection.AssemblyProductAttribute("WWPTools.WpfUI")]